Fixing Logging
