+++ lib/rpush.rb
@@ -13,7 +13,6 @@ require 'rpush/deprecation'
 require 'rpush/deprecatable'
 require 'rpush/logger'
 require 'rpush/multi_json_helper'
-require 'rpush/google_credential_cache'
 require 'rpush/configuration'
 require 'rpush/reflection_collection'
 require 'rpush/reflection_public_methods'
