+++ lib/rpush/daemon.rb
@@ -51,6 +51,7 @@ require 'rpush/daemon/apnsp8'
 
 require 'rpush/daemon/fcm/delivery'
 require 'rpush/daemon/fcm'
+require 'rpush/daemon/google_credential_cache'
 
 require 'rpush/daemon/gcm/delivery'
 require 'rpush/daemon/gcm'
